+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2020-07-25
+### Added
+- find by assembly 
+- find by type
+- [Add/Remove/Check] specific build target groups
+- [Add/Remove/Check] to all build target groups
+
+### Fixed 
+- assembly definition file only runs in editor
+
+### Changed
+- demo to be editor only
+- cleaned manifest
+
+### Removed
+- invalid namespace in unity package version
+- runtime demo
+
+
+
 ## [0.0.1] - 2020-05-22
 ### Added
 -Initial Commit